requirements.txt
# CATS Multi-Judge Evaluation Pipeline - Requirements
# Python 3.9+

# Core dependencies
httpx>=0.25.0                 # Async HTTP client for OpenRouter API
pyyaml>=6.0                   # Configuration file parsing
python-dotenv>=1.0.0          # Environment variable management

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# NLP and evaluation
nltk>=3.8.0                   # Text processing
scikit-learn>=1.3.0           # Metrics computation

# Async and concurrency
asyncio>=3.4.3
aiofiles>=23.0.0              # Async file operations
tqdm>=4.65.0                  # Progress bars

# Logging and monitoring
structlog>=23.1.0             # Structured logging

# Testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Original CATS dependencies (if using fallback features)
transformers>=4.30.0          # For tokenization
torch>=2.0.0                  # For local models (optional)

# Optional: for local NLI models
# sentence-transformers>=2.2.0
